change: change preinstall script

+++ debpkg/src/preinstall.sh
@@ -16,7 +16,7 @@ cleanInstall() {
 	fi
 
         if id "${vgstats_user}" >/dev/null 2>&1; then
-                echo 'user ${vgstats_user} already exists'
+                echo "user ${vgstats_user} already exists"
         else
                 if getent group "${vgstats_user}" >/dev/null 2>&1; then
                         useradd -p "*" -g "${vgstats_user}" -M "${vgstats_user}" -s /usr/sbin/nologin -d /nonexistent
@@ -30,8 +30,25 @@ cleanInstall() {
 upgrade() {
     	printf "\033[32m Pre Install of an upgrade\033[0m\n"
     	# Step 3(upgrade), do what you need
-        systemctl stop --all 'vgkeydesk@*.socket' 'vgkeydesk@*.service' ||:
-        systemctl stop --all 'vgstats@*.service' ||:
+        systemctl stop --all 'vgkeydesk@*.socket' ||:
+        
+        if [ -f /etc/systemd/system/vgstats@.service ]; then
+                systemctl stop --all 'vgstats@*.service' ||:
+                systemctl disable --all 'vgstats@*.service' ||:
+                rm -f /etc/systemd/system/vgstats@.service
+        fi
+
+        if [ -f /etc/systemd/system/vgkeydesk@.socket ]; then
+                systemctl stop --all 'vgkeydesk@*.socket' ||:
+                systemctl disable --all 'vgkeydesk@*.socket' ||:
+                rm -f /etc/systemd/system/vgkeydesk@.socket
+
+                find /etc/systemd/system/ -path "/etc/systemd/system/vgkeydesk@*.socket.d/*" -type f -name "listen.conf" -delete
+                find /etc/systemd/system/ -path "/etc/systemd/system/vgkeydesk@*.socket.d" -type d -empty -delete
+        fi
+
+        printf "\033[32m Reload the service unit from disk\033[0m\n"
+        systemctl daemon-reload ||:
 }
 
 # Step 2, check if this is a clean install or an upgrade
